+++ template/engine/include/trogue/trogue.hpp
@@ -6,6 +6,7 @@
 #include "trogue/config.hpp"
 #include "trogue/types.hpp"
 #include "trogue/coro.hpp"
+#include "trogue/task_runner.hpp"
 #include "trogue/scene.hpp"
 #include "trogue/terrain.hpp"
 #include "trogue/render.hpp"